+++ web/src/components/warning-select/index.tsx
@@ -3,93 +3,21 @@ import { ElementIcon, PeopleIcon, VehiclesIcon } from "@/icon/monitor";
 import { DownOutlined } from "@ant-design/icons";
 import { Checkbox, Switch } from "antd";
 import { clone } from "ramda";
-import { forwardRef, useImperativeHandle, useState } from "react";
+import { forwardRef, useImperativeHandle } from "react";
 import { Fragment } from "react/jsx-runtime";
-import { IItemProps, IWarningListProps, IWarningType } from "./props";
+import { IWarningType } from "./props";
+import { useAction } from "./hook";
 
 export const WarningSelect = forwardRef((_, ref) => {
-  const warningList: IWarningListProps[] = [
-    {
-      label: "人員",
-      value: IWarningType.People,
-      children: [
-        {
-          label: "人員檢測",
-          value: ICameraAiMonitorType.People,
-          defaultCheck: true,
-        },
-        {
-          label: "吸煙檢測",
-          value: ICameraAiMonitorType.Smoke,
-        },
-        {
-          label: "打架檢測",
-          value: ICameraAiMonitorType.Fight,
-        },
-        {
-          label: "安全配備檢測",
-          value: ICameraAiMonitorType.Costume,
-          children: [
-            {
-              label: "螢光衣",
-              value: ICameraAiMonitorType.FluorescentClothing,
-            },
-            {
-              label: "手套",
-              value: ICameraAiMonitorType.Gloves,
-            },
-            {
-              label: "安全鞋",
-              value: ICameraAiMonitorType.SafetyShoes,
-            },
-          ],
-        },
-      ],
-    },
-    {
-      label: "車輛",
-      value: IWarningType.Vehicles,
-      children: [
-        {
-          label: "識別車輛",
-          value: ICameraAiMonitorType.Vehicles,
-          defaultCheck: true,
-        },
-        {
-          label: "識別異常車輛",
-          value: ICameraAiMonitorType.AbnormalVehicles,
-        },
-      ],
-    },
-    {
-      label: "物體",
-      value: IWarningType.Element,
-      children: [
-        {
-          label: "動物入侵",
-          value: ICameraAiMonitorType.Animal,
-          children: [
-            {
-              label: "貓",
-              value: ICameraAiMonitorType.Cat,
-            },
-            {
-              label: "狗",
-              value: ICameraAiMonitorType.Dog,
-            },
-            {
-              label: "鳥",
-              value: ICameraAiMonitorType.Bird,
-            },
-          ],
-        },
-      ],
-    },
-  ];
-
-  const [selectValues, setSelectValues] = useState<ICameraAiMonitorType[]>([]);
-
-  const [checkIndex, setCheckIndex] = useState<IWarningType[]>([]);
+  const {
+    warningList,
+    selectValues,
+    checkIndex,
+    setSelectValues,
+    setCheckIndex,
+    handleSelect,
+    flatten,
+  } = useAction();
 
   useImperativeHandle(ref, () => ({
     selectValues,
@@ -97,76 +25,6 @@ export const WarningSelect = forwardRef((_, ref) => {
     setSelectValues,
     setCheckIndex,
   }));
-
-  const handleSelect = (type: number, subTypes?: number[]) => {
-    setSelectValues((prev) => {
-      let newData = clone(prev);
-
-      const animalData = [
-        ICameraAiMonitorType.Cat,
-        ICameraAiMonitorType.Dog,
-        ICameraAiMonitorType.Bird,
-      ];
-
-      const costumeData = [
-        ICameraAiMonitorType.FluorescentClothing,
-        ICameraAiMonitorType.Gloves,
-        ICameraAiMonitorType.SafetyShoes,
-      ];
-
-      if (subTypes) {
-        if (type === ICameraAiMonitorType.Costume) {
-          newData = newData.filter((value) => !costumeData.includes(value));
-        }
-
-        if (type === ICameraAiMonitorType.Animal) {
-          newData = newData.filter((value) => !animalData.includes(value));
-        }
-
-        newData = newData.concat(subTypes);
-
-        newData.push(type);
-
-        if (!subTypes.length) {
-          newData = newData.filter((value) => value !== type);
-        }
-
-        newData = newData.filter(
-          (value, index, self) => self.indexOf(value) === index
-        );
-      } else {
-        if (type === ICameraAiMonitorType.Animal) {
-          newData = newData.concat(animalData);
-        }
-
-        if (type === ICameraAiMonitorType.Costume) {
-          newData = newData.concat(costumeData);
-        }
-
-        const isExist = newData.findIndex((item) => item === type) !== -1;
-
-        if (isExist) {
-          newData = newData.filter((item) => item !== type);
-
-          if (type === ICameraAiMonitorType.Animal) {
-            newData = newData.filter((value) => !animalData.includes(value));
-          }
-
-          if (type === ICameraAiMonitorType.Costume) {
-            newData = newData.filter((value) => !costumeData.includes(value));
-          }
-        } else newData.push(type);
-      }
-
-      return newData;
-    });
-  };
-
-  const flatten = (items: IItemProps[]): IItemProps[] => {
-    return items.flatMap((item) =>
-      item.children ? [item, ...flatten(item.children)] : [item]
-    );
-  };
 
   const getIconHeader = (type: IWarningType | ICameraAiMonitorType) => {
     const iconProps = {
@@ -201,7 +59,7 @@ export const WarningSelect = forwardRef((_, ref) => {
   };
 
   return (
-    <div className="w-[25rem] bg-white rounded-[.5rem] pt-[1.3rem] px-[1.3rem]">
+    <div className="w-[25rem] bg-red-300 rounded-[.5rem] pt-[1.3rem] px-[1.3rem] overflow-y-auto">
       {warningList.map((item, index) => {
         return (
           <div className="mb-[2.5rem]" key={index}>
